refactor: localize sandbox snapshot tests
