+++ crates/sandbox-fc/src/snapshot/output.rs
@@ -138,3 +138,70 @@ pub(super) async fn snapshot_complete_marker_present(
         Err(e) => Err(e.into()),
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use crate::paths::SnapshotOutputPaths;
+
+    use super::*;
+
+    #[tokio::test]
+    async fn prepare_snapshot_output_removes_snapshot_artifacts_only() {
+        let dir = tempfile::tempdir().expect("tempdir");
+        let output = SnapshotOutputPaths::new(dir.path().to_path_buf());
+        let stale_work_file = output.work_dir().join("nested").join("stale.txt");
+        let unrelated = dir.path().join("keep.txt");
+
+        tokio::fs::create_dir_all(stale_work_file.parent().expect("parent"))
+            .await
+            .expect("create stale work dir");
+        tokio::fs::write(&stale_work_file, b"stale")
+            .await
+            .expect("write stale work file");
+        tokio::fs::write(&unrelated, b"keep")
+            .await
+            .expect("write unrelated file");
+        for artifact in [
+            output.snapshot(),
+            output.memory(),
+            output.cow(),
+            output.cow_bitmap(),
+            output.complete_marker(),
+        ] {
+            tokio::fs::write(&artifact, b"stale")
+                .await
+                .unwrap_or_else(|e| panic!("write {}: {e}", artifact.display()));
+        }
+
+        let work = prepare_snapshot_output(&output)
+            .await
+            .expect("prepare output");
+
+        assert_eq!(work, output.work_dir());
+        assert!(
+            tokio::fs::try_exists(output.work_dir()).await.unwrap(),
+            "work dir should be recreated"
+        );
+        assert!(
+            !tokio::fs::try_exists(stale_work_file).await.unwrap(),
+            "stale work contents should be removed"
+        );
+        for artifact in [
+            output.snapshot(),
+            output.memory(),
+            output.cow(),
+            output.cow_bitmap(),
+            output.complete_marker(),
+        ] {
+            assert!(
+                !tokio::fs::try_exists(&artifact).await.unwrap(),
+                "stale artifact should be removed: {}",
+                artifact.display()
+            );
+        }
+        assert!(
+            tokio::fs::try_exists(unrelated).await.unwrap(),
+            "non-snapshot output-dir contents should be preserved"
+        );
+    }
+}
